refactor: 边框9 tsx 改造
