+++ packages/docs/docs/Border/BorderBox9/demo.vue
@@ -1,9 +1,13 @@
 <template>
   <div demo-bg>
-    <dv-border-box-9>
+    <dv-border-box9>
       <div dv-bg>
         dv-border-box-9
       </div>
-    </dv-border-box-9>
+    </dv-border-box9>
   </div>
 </template>
+
+<script setup lang=ts>
+import { BorderBox9 as DvBorderBox9 } from '@kjgl77/datav-vue3'
+</script>
